Update README.md
You are to code an ExternalChainingHashMap, a key-value hash map with an external chaining collision resolution strategy. A HashMap maps unique keys to values and allows O(1) average case lookup of a value when the key is known.

The table should not contain duplicate keys, but can contain duplicate values. In the event of try-ing to add a duplicate key, replace the value in the existing (key, value) pair with the new value and return the old value.

You should implement two constructors for this HashMap. As per the javadocs, you should use con-structor chaining to implement the no-arg constructor.

Capacity

The starting capacity of the ExternalChainingHashMap using the default constructor should be the constant INITIAL CAPACITY defined in ExternalChainingHashMap.java. Reference the constant as-is. Do not simply copy the value of the constant. Do not change the constant. Do not regrow the backing array when removing elements.

If adding to the table would cause the load factor (LF) to exceed (greater than, not greater than or equal to) the max load factor constant provided in the java file, the table should be resized to have a capacity of 2n + 1, where n is the current capacity before adding the parameterized element. See the javadocs for specific instructions on when to resize. There is a method called resize that you should use for resizing.

Hash and Compression Functions